Rename datasetV2 folder
